move mocha tests to specs/ directory and load by a tiny loader
